updated parameter type for purchase location methods
Previously forgot to also change the parameter type after changing the variable type. This commit fixes that mistake.

+++ starbucks2-domain/src/java/model/Purchase.java
@@ -54,7 +54,7 @@ public class Purchase {
         return latPostion;
     }
 
-    public void setLatPostion(long latPostion) {
+    public void setLatPostion(double latPostion) {
         this.latPostion = latPostion;
     }
 
@@ -62,7 +62,7 @@ public class Purchase {
         return longPosition;
     }
 
-    public void setLongPosition(long longPosition) {
+    public void setLongPosition(double longPosition) {
         this.longPosition = longPosition;
     }
